+++ Source/GateRimSG1/GR_DefOf.cs
@@ -12,6 +12,7 @@ namespace GateRimSG1
         public static PawnKindDef SG1_TokraSymbiote;
         public static PawnKindDef SG1_TokraVoluntaryHost;
         public static FactionDef SG1_Tokra;
+        public static IncidentDef SG1_TokraPeacefulVisitors;
         public static JobDef SG1_GoauldAutonomousImplant;
         public static ThingDef SG1_GoauldRitualBasin;
         public static GeneDef SG1_JaffaLineage;